02-book-blog-agent/code/requirements.txt
anthropic>=0.40.0
openai>=1.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0
playwright>=1.40.0
Pillow>=10.0.0
pywin32>=306
